Workaround for wrong page numbers in longer TOC (#859)
MapFish print only. In some cases, the table of content page is longer than one page. This leads to wrong page numbers displayed.
This commit provides a workaround until a better solution is found. To enable it, set the multi_page_TOC property in your config to true.
This mode splits the table of contents into to separate pages. One for the available themes and another page for the remaining content of the table of content page.
This feature is disabled by default.